Split generators to files and strings
