+++ src/Package/Impl/Microsoft.VisualStudio.R.Package.csproj
@@ -271,12 +271,14 @@
     <Compile Include="Sql\Commands\ConfigurationSettingCommand.cs" />
     <Compile Include="Sql\ConnectionStringEditor.cs" />
     <Compile Include="Sql\Commands\AddDbConnectionCommand.cs" />
+    <Compile Include="Sql\DacPac\DacPacBuilder.cs" />
     <Compile Include="Sql\DragDrop\DataObject.cs" />
     <Compile Include="Sql\DragDrop\DropHandler.cs" />
     <Compile Include="Sql\DragDrop\DropHandlerProvider.cs" />
     <Compile Include="Sql\NativeMethods.cs" />
     <Compile Include="Sql\Publish\RCodePlacement.cs" />
-    <Compile Include="Sql\Publish\SProcGenerator.cs" />
+    <Compile Include="Sql\Publish\SProcScriptGenerator.cs" />
+    <Compile Include="Sql\Publish\SProcFileGenerator.cs" />
     <Compile Include="Sql\Extensions\SqlQuoteType.cs" />
     <Compile Include="Sql\Publish\SqlSProcPublishSettings.cs" />
     <Compile Include="Sql\Publish\SqlPublishDialogViewModel.cs" />
