+++ src/Package/Impl/Sql/Publish/SProcFileGenerator.cs
@@ -1,0 +1,86 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System.IO;
+using Microsoft.Common.Core.IO;
+
+namespace Microsoft.VisualStudio.R.Package.Sql.Publish {
+    internal sealed class SProcFileGenerator {
+        /// <summary>
+        /// Name of script file that contains SQL that creates R code table
+        /// </summary>
+        internal const string CreateRCodeTableScriptName = "CreateRCodeTable.sql";
+        /// <summary>
+        /// Name of the post-deployment script that inserts actual R code into the table
+        /// </summary>
+        internal const string PostDeploymentScriptName = "RCode.PostDeployment.sql";
+
+        private readonly IFileSystem _fs;
+
+        public SProcFileGenerator(IFileSystem fs) {
+            _fs = fs;
+        }
+
+        /// <summary>
+        /// Generates SQL scripts for the deployment of R code into SQL database.
+        /// Writes scripts to files and pushes files into the target database project.
+        /// </summary>
+        public void Generate(SqlSProcPublishSettings settings, EnvDTE.Project targetProject) {
+            var targetFolder = Path.Combine(Path.GetDirectoryName(targetProject.FullName), "R\\");
+            if (!_fs.DirectoryExists(targetFolder)) {
+                _fs.CreateDirectory(targetFolder);
+            }
+
+            var targetProjectItem = targetProject.ProjectItems.Item("R") ?? targetProject.ProjectItems.AddFolder("R");
+
+            if (settings.CodePlacement == RCodePlacement.Table) {
+                CreateRCodeTableFile(settings, targetProject, targetFolder, targetProjectItem);
+                CreatePostDeploymentScriptFile(settings, targetProject, targetFolder, targetProjectItem);
+            }
+            CreateStoredProcedureFiles(settings, targetProject, targetFolder, targetProjectItem);
+        }
+
+        /// <summary>
+        /// Create SQL file that defines table template that will hold R code
+        /// </summary>
+        internal void CreateRCodeTableFile(SqlSProcPublishSettings settings, EnvDTE.Project targetProject, string targetFolder, EnvDTE.ProjectItem targetProjectItem) {
+            var creatTableScriptFile = Path.Combine(targetFolder, CreateRCodeTableScriptName);
+
+            var g = new SProcScriptGenerator(_fs);
+            var script = g.CreateRCodeTableScript(settings);
+
+            _fs.WriteAllText(creatTableScriptFile, script);
+            targetProjectItem.ProjectItems.AddFromFile(creatTableScriptFile);
+        }
+
+        /// <summary>
+        /// Generates SQL post deployment script that pushes R code into a table
+        /// as well as 
+        /// </summary>
+        private void CreatePostDeploymentScriptFile(SqlSProcPublishSettings settings, EnvDTE.Project targetProject, string targetFolder, EnvDTE.ProjectItem targetProjectItem) {
+            var postDeploymentScript = Path.Combine(targetFolder, PostDeploymentScriptName);
+
+            var g = new SProcScriptGenerator(_fs);
+            var script = g.CreatePostDeploymentScript(settings);
+
+            _fs.WriteAllText(postDeploymentScript, script);
+
+            var item = targetProjectItem.ProjectItems.AddFromFile(postDeploymentScript);
+            item.Properties.Item("BuildAction").Value = "PostDeploy";
+        }
+
+        private void CreateStoredProcedureFiles(SqlSProcPublishSettings settings, EnvDTE.Project targetProject, string targetFolder, EnvDTE.ProjectItem targetProjectItem) {
+            var g = new SProcScriptGenerator(_fs);
+            var sprocs = g.CreateStoredProcedureScripts(settings);
+
+            foreach (var sprocName in sprocs.Keys) {
+                var template = sprocs[sprocName];
+                if (!string.IsNullOrEmpty(template)) {
+                    var sprocFile = Path.ChangeExtension(Path.Combine(targetFolder, sprocName), ".sql");
+                    _fs.WriteAllText(sprocFile, template);
+                    targetProjectItem.ProjectItems.AddFromFile(sprocFile);
+                }
+            }
+        }
+    }
+}
